+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.4.0] - 2022-08-24
+
 ## Breaking
 
 - `NewTestResourcePlugin` now receives a config struct as parameters with and will set safe defaults.
@@ -52,8 +54,8 @@
 - Github gist example.
 - Documentation.
 
-
-[unreleased]: https://github.com/slok/terraform-provider-goplugin/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/slok/terraform-provider-goplugin/compare/v0.4.0...HEAD
+[v0.4.0]: https://github.com/slok/terraform-provider-goplugin/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/slok/terraform-provider-goplugin/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slok/terraform-provider-goplugin/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/slok/terraform-provider-goplugin/compare/v0.1.0...v0.1.1